refactor: simplify the design by removing slice and mut
